Homework05

+++ HomeWork05/src/main/java/ru/otus/gpbu/pse/homework05/myybooks/dao/mappers/AuthorMapper.java
@@ -10,7 +10,7 @@ public class AuthorMapper implements RowMapper<Author> {
     @Override
     public Author mapRow(ResultSet resultSet, int i) throws SQLException {
 
-        int id = resultSet.getInt("id");
+        long id = resultSet.getLong("id");
         String name = resultSet.getString("name");
         return new Author(id, name);
     }
